slightly stricter type checking for OfxParser.parse
BeautifulSoup's behaviour when passed None varies depending on which
tree-building library is available, but in some cases it will return
a valid document instead of raising TypeError when given None.

This patch makes OfxParser.parse check for a file-like object, so that
it's not sensitive to the particular behaviour of BeautifulSoup (and the
caller gets a nicer exception as well).