check for <OFX> element when determining whether a file is empty
Similar to commit 8bbb4d10, BeautifulSoup's behaviour also varies when
passed an empty file depending on which parser is in use. It will either
return an empty string (this is what the code is currently expecting) or
a skeleton <html/> document with no contents. But what actually matters
for OfxParser.parse is whether it has the <OFX> root element.